Add ec script for opening Emacs app on OS X
If the server file doesn't exist, open the Emacs app. If it does exist,
delegate to emacslient (which was having problems finding the server
file, so pass it too). When '-n' is passed by $EDITOR to the open
command, it is processed as "open new instance", which is fine since
there is no instance. Otherwise it is passed to emacsclient such that
the call does not wait. This was necessary because the editor setting
for Git needs emacsclient to wait. I expect Git's use of editor to fail
if Emacs is not already open, as I am not passing -W for it to wait (and
would not want to either).