Export `SSH_AUTH_SOCK` whenever possible
Specifically if it is not already set but a likely file exists, then set
it. This enables other applications to use a forwarded SSH key even if
they did not inherit the SSH session environment (such as Emacs started
by systemd, when taught to look for the variable from the shell, which
is also why it must happen before the short-circuit for dumb terms).